feat: move helper function
